refact mattermost structure
