+++ spec/workers/integration_worker_spec.rb
@@ -4,7 +4,7 @@ describe IntegrationWorker do
   let(:integration) { create(:integration) }
 
   it 'should send a message to mattermost' do
-    expect(Mattermost).to receive(:send).with(
+    expect(Integrations::Mattermost::Service).to receive(:send).with(
       integration.data['private_uri'],
       integration.data['channel'],
       integration.data['bot_username'],
@@ -15,7 +15,7 @@ describe IntegrationWorker do
 
   it 'should read URI from ENV' do
     integration.data['private_uri'] = 'INTEGRATION_URI_MATTERMOST'
-    expect(Mattermost).to receive(:send).with(
+    expect(Integrations::Mattermost::Service).to receive(:send).with(
       'http://foo.com',
       integration.data['channel'],
       integration.data['bot_username'],
